Move Server Message classes to separate file
