+++ triggers/service/src/main/scala/com/digitalasset/daml/lf/engine/trigger/ServiceMain.scala
@@ -46,22 +46,6 @@ import java.time.LocalDateTime
 
 import com.daml.lf.engine.trigger.dao._
 
-sealed trait Message
-final case class GetServerBinding(replyTo: ActorRef[ServerBinding]) extends Message
-case object Stop extends Message
-final case class StartFailed(cause: Throwable) extends Message
-final case class Started(binding: ServerBinding) extends Message
-final case class TriggerStarting(runningTrigger: RunningTrigger) extends Message
-final case class TriggerStarted(runningTrigger: RunningTrigger) extends Message
-final case class TriggerInitializationFailure(
-    runningTrigger: RunningTrigger,
-    cause: String
-) extends Message
-final case class TriggerRuntimeFailure(
-    runningTrigger: RunningTrigger,
-    cause: String
-) extends Message
-
 class Server(
     ledgerConfig: LedgerConfig,
     runnerConfig: TriggerRunnerConfig,
